build: fix compiler warnings and transition to C11
- Remove unused parameters from internal bitstream.c functions and
  suppress unused-parameter warning in signal_handler callback.
- Fix negative bitwise shift in WriteTNSData.
- Resolve sign-comparison and missing-initializer warnings in frame.c,
  main.c, and input.c with casts and explicit initializers.
- Use correct %u format specifiers for unsigned track/disc numbers in
  main.c.
- Use MinBitrate(void) instead of an empty K&R-style prototype to
  satisfy -Wstrict-prototypes.
- Update meson.build to use c_std=gnu11,c11 and warning_level=3.
- Standardize on <string.h> in libfaac/util.h and add fallback
  definitions for M_PI and M_SQRT2 for portability across compilers.

Verified clean build with -Wall -Wextra -Wpedantic at warning_level=3
(zero warnings).

+++ frontend/input.c
@@ -195,7 +195,7 @@ pcmfile_t *wav_open_read(const char *name, int rawinput)
     if (fmtsize < 16)
 	return NULL;
 
-   if (fread(&wave, 1, fmtsize, wave_f) != fmtsize)
+   if (fread(&wave, 1, fmtsize, wave_f) != (size_t)fmtsize)
         return NULL;
 
     seekcur(wave_f, riffsub.len - fmtsize);
